Split x64_rip_rel_abs.py into 3 separate unit tests.
- x64_lea_label.py tests label addressing on x64 using all syntaxes (this
  demonstrates the incorrect addressing of labels on x64).
- x64_lea_label_rel.py tests rip relative label addressing on x64 using all
  available syntaxes (this demonstrates nasm syntax not being parsed).
- x64_lea_label_nasm_abs.py tests explicit absolute addressing on x64
  using nasm syntax (no equivalent exists in other syntaxes afaik).
